requirements.txt
Django==2.2.22
django-filter==2.4.0
django-tenant-schemas==1.10.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
psycopg2-binary==2.8.6
requests==2.22.0
six==1.16.0